split out limit_contract_functions_number tests
